feat: add option to enforce unique attribute
Enforcing Uniqueness on an Index can go a long way since higher-level
application code doesn't need to add explicit checks for it.

To keep the change backwards compatible, changes have been made to
`Insert()` method to respect Unique constraints.

Insert will not allow update if `EnforceUniqueness` is set to
true on the id schema.
If any other schema of the table has EnforceUnique, updates using
`Insert()` will error out if the same value exists in the schema.

A new `Update()` has been added to allow for updates when `EnforceUniqueness`
is set.