Document correct usage of a ResultIterator
We receicently had [a panic in Consul](https://github.com/hashicorp/consul/issues/9566)
that was caused by incorrect usage of a ResultIterator. A minimal
reproduction of this panic can be seen [here](https://github.com/hashicorp/go-memdb/compare/panic-delete-inside-iter).

Many of us learned that there are undocumented rules for using a
ResultIterator correct with a write transaction.

This commit attempts to document that safe use of a ResultIterator.