Panic when ResultIterator is used incorrectly.
Previously this could panic anyway (see https://github.com/hashicorp/go-memdb/pull/85),
but only when the nodse in an index line up in a very specify way. This
meant that tests may not exercise the bug, and the panic would get into
a release.

By adding an explicit panic here we ensure that any basic test coverage
of an iteration would fail.

In the case of inserts it would not panic, but could return incorrect
results (most likely duplicate). This panic ensures we fail instead of
returning incorrect results.

It would have been nice to error instead of panic, but there's no place
in the current interface to return an error.